Gmail API commit
Have not worked on the front end as I guess the front end for this part is already designed. I have just used a simple HTML page for the front end which is the "form.js" file.  This code on its own wont run as we have to provide a client ID, client secret, API key and a redirect URI which will be specific to each account.